Take reducers and actions out of App component
